backend/requirements.txt
flask==2.3.3
flask-sqlalchemy==3.1.1
flask-marshmallow==0.15.0
marshmallow-sqlalchemy==0.29.0
flask-migrate==4.0.5
flask-restful==0.3.10
flask-cors==4.0.0
python-dotenv==1.0.0
